+++ purchasely_android_player/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 5.1.1
+### Android
+- Fix image display issue between default and selected state
+- Add `Purchasely.closeAllScreens()` method with possibility to intercept action close_all
+## iOS
+- TODO
 ## 5.1.0
 ## 🚀 Campaigns Support
 Introducing our powerful new Campaigns feature, enabling you to launch targeted in-app campaigns for specific audiences or all users right at app launch.